+++ R/latch.R
@@ -3,7 +3,7 @@
 #' Latch an error or a warning to an object
 #' to indicate an issue associated with it.
 #' These can later be checked with [is.e()]
-#' and [is.w()], and can also be [enforce()].
+#' and [is.w()], and can also be [resolve()].
 #' 
 #' @param obj Object to latch the error or 
 #' warning onto.
@@ -21,7 +21,7 @@
 #' is.e(problematic)
 #' 
 #' do_sth_with_x <- function(x){
-#'  enforce(x)
+#'  resolve(x)
 #'  x + 1
 #' }
 #'