+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Rename the package from err to erratum: err is already on CRAN.
 - Rename `latche`, and `latchw` to `latch.e`, and `latch.w`.
+- Add `return` method for error escalation.
+
 # erratum 1.0.1
 
 - Add `latche`, and `latchw` to latch errors and warnings onto object to later `enforce` them.